✨ Initial <Dialog> component and tests
Where possible this component follows the reference implementation set out by https://headlessui.dev/vue/dialog

However, this is not a full implementation. This cut of the component
does not include:

1. Implementation of the `static` prop
2. Any consideration around css transitions
3. Nesting of Dialogs

Also, I have not done a thorough once over of the reference
implementation API to ensure this is as close to it as possible. That
will come in a subsequent PR.