fix(menu): pass selector to `focus-trap`
Previously, we were passing an element to the focus-trap options. This worked fine in pre-4.0.0 Ember, but the timing has changed in Ember 4.0; the getter is being evaluated before the DOM node has been inserted, rather than after. This means that the getter returns `null` rather than the element.

By passing the selector, we allow the `focus-trap` library to determine when we actually perform the query for the DOM node. It does this after the element has been inserted into the DOM instead!