chore(ci): disable `beta` and `canary` test jobs
These Ember builds seem to be really unstable; at the very least, `ember-element-helper`'s tests are failing on these Ember versions, so we have little chance of the tests passing in this repo.

For now, we should just disable these tests since we shouldn't allow the failing tests on unreleased Ember versions to falsely indicate that the tests are not passing under normal circumstances.

We *could* run the tests and allow them to fail for these versions, but falsely showing that the jobs pass seems as helpful as just not running them at all.