Some progress with failing tests
Most of the problems come from setting `this.activeOptionIndex` in
`combobox.js` (and in `listbox.js`) in lots of places and also reading
("consuming") it various places which I think triggers the error.

I wonder if `this.activeOptionIndex` can be made into a getter, maybe
using a `this._activeOptionIndex` for the cases where it really needs to
be explicitly set. (Or maybe `activeOptionGuid` is already used for
that).