Add Combobox component (#165)
* Add Combobox component

Co-authored-by: David McNamara <david.mcnamara@phorest.com>

This was really done by David but the file structure of the add-on has been
drastically changed so the feasible way was to copy his work from the
previous place. All props should go to him.

* Fix linting errors

* Replace did-insert and will-destroy modifiers

* Some progress with failing tests

Most of the problems come from setting `this.activeOptionIndex` in
`combobox.js` (and in `listbox.js`) in lots of places and also reading
("consuming") it various places which I think triggers the error.

I wonder if `this.activeOptionIndex` can be made into a getter, maybe
using a `this._activeOptionIndex` for the cases where it really needs to
be explicitly set. (Or maybe `activeOptionGuid` is already used for
that).

* Try to make this work with `activeOptionGuid` not consuming `activeOptionIndex`

* [WIP] Use activeOptionGuid to determine option activeness

Also, break the inheritance link between Listbox and Combobox.

* Untrack optionElements

* Fix setting value when clicking an option

* Restore tracking optionElements

* Fix the usage of reverse()

* Use run.next to sidestep the "mutation after consumption" error for optionsElement

* Make going to the previous and next elements work

* Current implementation doesn't go around the options when navigating with keyboard

* Unset the active option on mouse click

* Register the button, too

* Have Enter select the active option

* Remove comment about strategy

* Refactor "walkers" and add `setLastOptionActive`

* Revert changes in Listbox

* Fix using up arrow to cycle through combobox options

* Fix (de)selecting combobox option for multi-select

* Guard against lacking an activeOption when walking

* Add links to Combobox pages in test app

* Remove unused activeOptionIndex

* Fix selecting the first option on Enter